Changed: Moved `canonicalize_cli_path` to `util` function.
